Use explicit pointer casts for future C++ compatibility

+++ types.h
@@ -398,7 +398,7 @@ typedef struct levellist_entry
   int nphixstargets;                         /// length of phixstargets array:
   int stat_weight;                           /// Statistical weight of this level.
 
-  long cont_index;                            /// Index of the continuum associated to this level. Negative number.
+  int cont_index;                            /// Index of the continuum associated to this level. Negative number.
 #if (!NO_LUT_PHOTOION || !NO_LUT_BFHEATING)
   int closestgroundlevelcont;
 #endif
